Taking FaultException out of GbxRemote file
